@@ -69,7 +69,7 @@
     "@babel/plugin-transform-runtime": "7.7.4",
     "@babel/preset-env": "7.8.7",
     "@babel/preset-react": "7.8.3",
-    "@babel/preset-typescript": "7.8.3",
+    "@babel/preset-typescript": "7.10.1",
     "@epegzz/sass-vars-loader": "5.0.0",
     "@types/classnames": "2.2.9",
     "@types/i18n-js": "3.0.1",
@@ -130,7 +130,7 @@
     "sass-loader": "8.0.0",
     "serve": "11.2.0",
     "style-loader": "1.0.1",
-    "typescript": "3.7.5",
+    "typescript": "3.9.5",
     "url-loader": "3.0.0",
     "webpack": "4.41.2",
     "webpack-cli": "3.3.10",
