@@ -1,6 +1,4 @@
 import * as React from 'react'
-import { ThemeProvider } from '@material-ui/styles'
-import { Theme } from 'themes/types'
 
 import { config } from 'constants/global'
 import Navigation from 'components/Navigation'
@@ -12,7 +10,7 @@ import { AppProps } from './types'
 import { initTabs, loginTabs, gobackConfig } from './config'
 import NotificationStack from 'containers/NotificationStack'
 
-const theme: Theme = require(`themes/${process.env.THEME || 'default'}`).default
+import logo from 'theme/images/logo.png'
 
 const App: React.FC<AppProps> = ({ auth, history, loginUser, logout }) => {
   const [currentTab, setCurrentTab] = React.useState(0)
@@ -30,7 +28,7 @@ const App: React.FC<AppProps> = ({ auth, history, loginUser, logout }) => {
       loginUser({ token: auth.authToken })
     }
 
-    if (pathname.startsWith('/dashboard/')) {
+    if (pathname.startsWith('/dashboard/') || pathname.startsWith('/profile')) {
       setNavScrolled(true)
     } else {
       setNavScrolled(false)
@@ -38,7 +36,10 @@ const App: React.FC<AppProps> = ({ auth, history, loginUser, logout }) => {
 
     const gb = gobackConfig.find((item) => pathname.includes(item.path))
 
-    if (pathname.startsWith('/auth')) {
+    if (pathname.startsWith('/auth') ||
+    pathname.startsWith('/confirmation') ||
+    pathname.startsWith('/registration') ||
+    pathname.startsWith('/forgot')) {
       setNavigations(false)
     } else {
       setNavigations(true)
@@ -81,7 +82,7 @@ const App: React.FC<AppProps> = ({ auth, history, loginUser, logout }) => {
   }, [auth.user])
 
   return (
-    <ThemeProvider theme={theme}>
+    <div style={{ height: '100%' }}>
       {navigations &&
         <Navigation
           key='app-navigation'
@@ -91,8 +92,8 @@ const App: React.FC<AppProps> = ({ auth, history, loginUser, logout }) => {
           subTabIndex={currentSubTab}
           onTabChange={handleOnTabChange}
           onSubTabChange={handleOnSubTabChange}
-          name={config.navbar.name}
-          logoSrc={theme.images.apiLogo}
+          name={config.portalName}
+          logoSrc={logo}
           user={auth.user}
           forceScrolled={navScrolled}
           showBackButton={gobackLabel.length > 0}
@@ -108,7 +109,7 @@ const App: React.FC<AppProps> = ({ auth, history, loginUser, logout }) => {
       <InformDialog />
 
       <NotificationStack />
-    </ThemeProvider>
+    </div>
   )
 }
 
