@@ -1,8 +1,7 @@
 import { makeStyles } from '@material-ui/styles'
-import { Theme } from 'themes/types'
-import { colorPicker } from 'util/colorPicker'
+import { config } from 'constants/global'
 
-export default makeStyles((theme: Theme) => ({
+export default makeStyles({
   cardContainer: {
     display: 'flex',
     flexDirection: 'row',
@@ -10,13 +9,13 @@ export default makeStyles((theme: Theme) => ({
     justifyContent: 'space-between',
     width: '100%',
     height: 42,
-    backgroundColor: theme.palette.background.default,
+    backgroundColor: config.palette.background.default,
     paddingLeft: 10,
     borderBottomWidth: 0,
     borderTopWidth: 1,
     borderLeftWidth: 0,
     borderRightWidth: 0,
-    borderColor: colorPicker(theme.palette.grey, 300, '#D1D1D1'),
+    borderColor: config.palette.greyScales[300],
     borderStyle: 'solid',
   },
   options: {
@@ -30,11 +29,11 @@ export default makeStyles((theme: Theme) => ({
   },
   apiVersion: {
     width: '25%',
-    color: colorPicker(theme.palette.grey, 400, '#ACACAC'),
+    color: config.palette.greyScales[400],
   },
   apiApps: {
     display: 'flex',
     flexDirection: 'row',
     width: '25%',
   },
-}))
+})
