@@ -4,8 +4,10 @@ import virtualize from 'react-swipeable-views-utils/lib/virtualize'
 import bindKeyboard from 'react-swipeable-views-utils/lib/bindKeyboard'
 import SwipeableViews from 'react-swipeable-views'
 import Dots from 'material-ui-dots'
-import { makeStyles } from '@material-ui/styles'
 import { CarouselProps, RendererProps } from './types'
+import useStyles from './styles'
+import clsx from 'clsx'
+import { useTranslation } from 'react-i18next'
 
 const VirtualizeSwipeViews = bindKeyboard(virtualize(SwipeableViews))
 const VirtualizeAutoPlaySwipeViews = autoPlay(VirtualizeSwipeViews)
@@ -16,23 +18,53 @@ const modulo = (a: number, n: number) => ((a % n) + n) % n
 const carouselSlideRenderer = (children: React.ReactNodeArray) =>
   ({ index, key }: RendererProps) => React.cloneElement(children[modulo(index, children.length)] as any, { key })
 
-const useStyles = makeStyles({
-  root: {
-    width: '100%',
-  },
-  dots: {
-    paddingTop: 36,
-    margin: '0 auto',
-  },
-})
+const renderContent = (slidesConfig: any, classes: any, t: any) => (
+  slidesConfig.map((slide: any, i: number) => (
+    <div key={i} className={classes.slide}>
+      <img className={classes.slideImage} src={slide.img} alt='' />
 
-const Carousel: React.FC<CarouselProps> = ({ children, autoplay, ...other }) => {
-  if (children == null) {
-    return null
-  }
+      <div className={classes.spacer} />
 
-  const slideRenderer = carouselSlideRenderer(children)
+      <section className={classes.slideInfo}>
+        <h1 className={classes.slideInfoH1}>{t(slide.title)}</h1>
+
+        {/** TODO: strings are not being escaped from translations, this needs to be reviwed */}
+        <p className={classes.slideInfoParagraph} dangerouslySetInnerHTML={{ __html: t(slide.subtitle) }} />
+
+        {slide.content && (
+          <p
+            className={classes.slideInfoParagraph}
+            dangerouslySetInnerHTML={{ __html: t(slide.content) }}
+          />
+        )}
+
+        <div
+          role='button'
+          arial-label='register'
+          className={clsx(classes.btn, {
+            [classes.btn2]: slide.btn === 2,
+            [classes.btn3]: slide.btn === 3,
+          })}
+        >
+          {!slide.disabled &&
+            <a href={slide.link} className={classes.buttonLink}>
+              {t(slide.button)}
+            </a>}
+          {slide.disabled &&
+            <div>{t(slide.button)}</div>}
+        </div>
+      </section>
+    </div>
+  ))
+)
+
+const Carousel: React.FC<CarouselProps> = ({ slideConfig }) => {
   const classes = useStyles()
+  const [t] = useTranslation()
+
+  const autoplay = true
+  const content = renderContent(slideConfig, classes, t)
+  const slideRenderer = carouselSlideRenderer(content)
   const [slideIndex, setSlideIndex] = React.useState(0)
 
   function handleChange (slideIndex: number) {
@@ -41,14 +73,13 @@ const Carousel: React.FC<CarouselProps> = ({ children, autoplay, ...other }) => 
 
   const Views = autoplay ? (
     <VirtualizeAutoPlaySwipeViews
-      {...other}
+      interval={8000}
       slideRenderer={slideRenderer}
       index={slideIndex}
       onChangeIndex={handleChange}
     />
   ) : (
     <VirtualizeSwipeViews
-      {...other}
       slideRenderer={slideRenderer}
       index={slideIndex}
       onChangeIndex={handleChange}
@@ -60,8 +91,8 @@ const Carousel: React.FC<CarouselProps> = ({ children, autoplay, ...other }) => 
       {Views}
 
       <Dots
-        count={children.length}
-        index={modulo(slideIndex, children.length)}
+        count={content.length}
+        index={modulo(slideIndex, content.length)}
         className={classes.dots}
         onDotClick={handleChange}
       />
