@@ -1,7 +1,5 @@
 export interface CarouselProps extends Omit<React.HTMLAttributes<HTMLDivElement>, 'children'> {
-  children: React.ReactNodeArray,
-  autoplay?: boolean,
-  interval?: number,
+  slideConfig: any,
 }
 
 export type RendererProps = {
