@@ -1,2 +1,9 @@
 
-export type PanelProps = React.HTMLAttributes<HTMLElement>
+export type PanelProps = {
+  className?: string,
+  title?: string,
+  subtitle?: string,
+  content?: string,
+  button?: string,
+  link?: string,
+}
