@@ -1,8 +1,7 @@
 import { makeStyles } from '@material-ui/styles'
-import { Theme } from 'themes/types'
-import { colorPicker } from 'util/colorPicker'
+import { config } from 'constants/global'
 
-export default makeStyles((theme: Theme) => ({
+export default makeStyles(({
   table: {
     display: 'flex',
     flexDirection: 'column',
@@ -10,11 +9,11 @@ export default makeStyles((theme: Theme) => ({
     justifyContent: 'center',
     width: '100%',
     cursor: 'pointer',
-    borderRadius: theme.shape.borderRadius,
+    borderRadius: config.dimensions.borderRadius,
     backgroundColor: 'white',
     overflow: 'hidden',
     borderStyle: 'solid',
-    borderColor: colorPicker(theme.palette.grey, 400, '#D1D1D1'),
+    borderColor: config.palette.greyScales[300],
     border: 1,
     borderTopWidth: 0,
   },
@@ -26,7 +25,7 @@ export default makeStyles((theme: Theme) => ({
     justifyContent: 'space-between',
     height: 42,
     paddingLeft: 10,
-    color: colorPicker(theme.palette.grey, 400, '#ACACAC'),
+    color: config.palette.greyScales[400],
   },
   actions: {
     paddingRight: 30,
@@ -45,7 +44,7 @@ export default makeStyles((theme: Theme) => ({
     borderBottomWidth: 0,
     borderRightWidth: 0,
     borderLeftWidth: 0,
-    borderColor: colorPicker(theme.palette.grey, 400, '#D1D1D1'),
+    borderColor: config.palette.greyScales[300],
   },
   apiContainer: {
     width: '100%',
@@ -69,7 +68,7 @@ export default makeStyles((theme: Theme) => ({
     borderBottomWidth: 0,
     borderRightWidth: 1,
     borderLeftWidth: 1,
-    borderColor: colorPicker(theme.palette.grey, 400, '#D1D1D1'),
+    borderColor: config.palette.greyScales[300],
   },
   appsListCardContainer: {
     display: 'flex',
@@ -78,12 +77,12 @@ export default makeStyles((theme: Theme) => ({
     width: '100%',
     minHeight: 42,
     borderStyle: 'solid',
-    borderRadius: theme.shape.borderRadius,
+    borderRadius: config.dimensions.borderRadius,
     borderTopWidth: 1,
     borderBottomWidth: 1,
     borderRightWidth: 1,
     borderLeftWidth: 1,
-    borderColor: colorPicker(theme.palette.grey, 400, '#D1D1D1'),
+    borderColor: config.palette.greyScales[300],
   },
   icons: {
     display: 'flex',
@@ -91,7 +90,7 @@ export default makeStyles((theme: Theme) => ({
     justifyContent: 'center',
     width: 42,
     height: 42,
-    color: theme.palette.text.primary,
+    color: config.palette.text.primary,
   },
   cardContainer: {
     display: 'flex',
@@ -100,9 +99,9 @@ export default makeStyles((theme: Theme) => ({
     minHeight: 438,
     width: 289,
     backgroundColor: 'white',
-    borderRadius: theme.shape.borderRadius,
+    borderRadius: config.dimensions.borderRadius,
     borderStyle: 'solid',
-    borderColor: colorPicker(theme.palette.grey, 400, '#D1D1D1'),
+    borderColor: config.palette.greyScales[300],
     borderWidth: 1,
     overflow: 'hidden',
   },
@@ -130,13 +129,13 @@ export default makeStyles((theme: Theme) => ({
     width: '100%',
     height: 36,
     padding: 15,
-    backgroundColor: theme.palette.background.white,
+    backgroundColor: '#fff',
     borderTopWidth: 1,
     borderBottomWidth: 0,
     borderLeftWidth: 0,
     borderRightWidth: 0,
     borderStyle: 'solid',
-    borderColor: colorPicker(theme.palette.grey, 400, '#D1D1D1'),
+    borderColor: config.palette.greyScales[300],
   },
   viewRow: {
     display: 'flex',
@@ -150,8 +149,8 @@ export default makeStyles((theme: Theme) => ({
     alignItems: 'center',
     marginBottom: 20,
     borderWidth: 1,
-    borderRadius: theme.shape.borderRadius,
-    borderColor: colorPicker(theme.palette.grey, 400, '#D1D1D1'),
+    borderRadius: config.dimensions.borderRadius,
+    borderColor: config.palette.greyScales[300],
     borderStyle: 'solid',
     cursor: 'pointer',
     overflow: 'hidden',
@@ -166,13 +165,13 @@ export default makeStyles((theme: Theme) => ({
     justifyContent: 'flex-start',
     width: 200,
     padding: 10,
-    color: colorPicker(theme.palette.grey, 400, '#ACACAC'),
-    backgroundColor: theme.palette.background.white,
+    color: config.palette.greyScales[400],
+    backgroundColor: '#fff',
     borderStyle: 'none',
   },
   icon: {
-    color: colorPicker(theme.palette.grey, 400, '#ACACAC'),
-    backgroundColor: theme.palette.background.white,
+    color: config.palette.greyScales[400],
+    backgroundColor: '#fff',
     height: '100%',
     cursor: 'pointer',
   },
@@ -182,13 +181,13 @@ export default makeStyles((theme: Theme) => ({
     justifyContent: 'center',
     width: 42,
     height: 42,
-    color: theme.palette.text.primary,
+    color: config.palette.text.primary,
     backgroundColor: 'white',
     borderLeftWidth: 1,
     borderTopWidth: 0,
     borderBottomWidth: 0,
     borderRightWidth: 0,
-    borderColor: colorPicker(theme.palette.grey, 400, '#D1D1D1'),
+    borderColor: config.palette.greyScales[300],
     borderStyle: 'solid',
     cursor: 'pointer',
   },
@@ -196,6 +195,6 @@ export default makeStyles((theme: Theme) => ({
     width: '100%',
   },
   vNumber: {
-    color: colorPicker(theme.palette.grey, 400, '#ACACAC'),
+    color: config.palette.greyScales[400],
   },
 }))
