@@ -2,13 +2,15 @@ import { LocaleOption } from 'language/types'
 
 export interface SandboxConfig {
   includes: { [prop: string]: boolean },
-  navbar: {
-    name: string,
-    logoUrl: string,
-  },
+  portalName: string,
+  clientName: string,
+  infra: any,
+  social: any,
   footer: {
     copyright: string,
-    logoUrl: string,
   },
   i18n: LocaleOption[],
+  palette: any,
+  dimensions: any,
+  pages: any,
 }
