@@ -56,18 +56,21 @@ export function * createApp (action: CreateAppAction) {
       'user_id': action.appData.userId,
       subscriptions: action.appData.subscriptions,
       // TODO: change
-      'pub_urls': [{
-        url: 'http://ssss.com',
-        type: 'client',
-      }],
+      // 'pub_urls': [{
+      //   url: 'http://ssss.com',
+      //   type: 'client',
+      // }],
     }
-
+    const accessToken = yield select(
+      (state: Store) => state.auth.authToken)
     const createAppUrl = `${API_URL}${SIGNUP_PORT}/app/create`
+
     yield call(request, {
       url: createAppUrl,
       method: 'POST',
       headers: {
         'content-type': 'application/x-www-form-urlencoded',
+        'x-access-token': accessToken,
       },
       data: qs.stringify(data),
     })
@@ -90,13 +93,16 @@ export function * updateApp (action: UpdateAppAction) {
       'sandbox_id': '1',
       'pub_urls': [action.appData.pubUrls],
     }
-
+    const accessToken = yield select(
+      (state: Store) => state.auth.authToken)
     const updateAppUrl = `${API_URL}${SIGNUP_PORT}/app/update/${action.appData.appId}`
+
     yield call(request, {
       url: updateAppUrl,
       method: 'PUT',
       headers: {
         'content-type': 'application/x-www-form-urlencoded',
+        'x-access-token': accessToken,
       },
       data: qs.stringify(data),
     })
@@ -122,12 +128,14 @@ export function * updateApp (action: UpdateAppAction) {
 export function * getUsersApps (action: GetUserAppsAction) {
   try {
     const getUserAppsUrl = `${API_URL}${SIGNUP_PORT}/app/list/${action.userId}`
-
+    const accessToken = yield select(
+      (state: Store) => state.auth.authToken)
     const response = yield call(request, {
       url: getUserAppsUrl,
       method: 'GET',
       headers: {
         'content-type': 'application/x-www-form-urlencoded',
+        'x-access-token': accessToken,
       },
     })
 
@@ -157,11 +165,15 @@ export function * getUsersApps (action: GetUserAppsAction) {
 export function * deleteApp (action: DeleteAppAction) {
   try {
     const deleteAppUrl = `${API_URL}${SIGNUP_PORT}/app/delete/${action.appId}`
+    const accessToken = yield select(
+      (state: Store) => state.auth.authToken)
+
     yield call(request, {
       url: deleteAppUrl,
       method: 'DELETE',
       headers: {
         'content-type': 'application/x-www-form-urlencoded',
+        'x-access-token': accessToken,
       },
     })
     yield put(deleteAppSuccess())
@@ -176,12 +188,15 @@ export function * deleteApp (action: DeleteAppAction) {
 
 export function * getAppDetails (action: GetAppDetails) {
   const getUserAppsUrl = `${API_URL}${SIGNUP_PORT}/app/list/${action.userId}`
+  const accessToken = yield select(
+    (state: Store) => state.auth.authToken)
 
   const response = yield call(request, {
     url: getUserAppsUrl,
     method: 'GET',
     headers: {
       'content-type': 'application/x-www-form-urlencoded',
+      'x-access-token': accessToken,
     },
   })
 
@@ -199,8 +214,8 @@ export function * getAppDetails (action: GetAppDetails) {
       enable: userApp.enable,
       createdAt: userApp.createdAt,
       updatedAt: userApp.updatedAt,
-      clientId: userApp.client_data.client_id,
-      clientSecret: userApp.client_data.client_secret,
+      clientId: userApp.client_data.client.client_id,
+      clientSecret: userApp.client_data.client.client_secret,
     }
   ))
 
@@ -214,6 +229,8 @@ export function * addAppSubscriptionSaga (action: AddAppSubscriptionAction) {
   const userApps: AppData[] = yield select((store: Store) => store.applications.userApps)
   const appInfo: AppData[] = userApps.filter(app => app.appId === action.appId)
   const appSubscriptions = appInfo[0].subscriptions.map((sub: Api) => sub.id)
+  const accessToken = yield select(
+    (state: Store) => state.auth.authToken)
 
   for (const apiIndx in apis) {
     if (apis[apiIndx].name === action.apiName) appSubscriptions.push(apis[apiIndx].id)
@@ -230,6 +247,7 @@ export function * addAppSubscriptionSaga (action: AddAppSubscriptionAction) {
       method: 'PUT',
       headers: {
         'content-type': 'application/x-www-form-urlencoded',
+        'x-access-token': accessToken,
       },
       data: qs.stringify(data),
     })
@@ -265,6 +283,8 @@ export function * removeAppSubscriptionSaga (action: RemoveAppSubscriptionAction
   const appInfo: AppData[] = userApps.filter(app => app.appId === action.appId)
   const appSubscriptions = appInfo[0].subscriptions.map((sub: Api) => sub.id)
   const subscriptionsToRemove = apis.filter((api: Api) => api.name === action.apiName).map((api: Api) => api.id)
+  const accessToken = yield select(
+    (state: Store) => state.auth.authToken)
 
   const data = {
     subscriptions: appSubscriptions.filter(apiId => !subscriptionsToRemove.includes(apiId)),
@@ -277,6 +297,7 @@ export function * removeAppSubscriptionSaga (action: RemoveAppSubscriptionAction
       method: 'PUT',
       headers: {
         'content-type': 'application/x-www-form-urlencoded',
+        'x-access-token': accessToken,
       },
       data: qs.stringify(data),
     })
