@@ -2,7 +2,11 @@ import * as React from 'react'
 import { RequireAuthProps } from './types'
 import { Redirect } from 'react-router'
 
-export const RequireAuth: React.FC<RequireAuthProps> = ({ auth, component: Component, ...rest }) => {
+export const RequireAuth: React.FC<RequireAuthProps> = ({
+  auth,
+  component: Component,
+  ...rest
+}) => {
   if (auth) {
     if (auth.user && !auth.isAuthorizing) {
       return <Component {...rest} />
