@@ -1,8 +1,25 @@
-import { put, call, takeLatest } from 'redux-saga/effects'
+import {
+  put,
+  call,
+  takeLatest,
+} from 'redux-saga/effects'
 import request from 'util/request'
-import { authActions, LOGIN, LOGIN_USER, LOGIN_SUCCESS } from './ducks'
-import { AUTH_URL, LOGIN_PORT } from 'constants/endpoints'
+import {
+  authActions,
+  LOGIN,
+  LOGIN_USER,
+  LOGIN_SUCCESS,
+  FORGOT_PASSWORD,
+  RECOVER_PASSWORD,
+} from './ducks'
+import {
+  AUTH_URL,
+  LOGIN_PORT,
+  API_URL,
+} from 'constants/endpoints'
+import { roleNameOptions } from 'containers/Profile/types'
 import qs from 'qs'
+import { openNotification } from 'containers/NotificationStack/ducks'
 
 import { AnyAction } from 'redux'
 
@@ -44,9 +61,8 @@ function * loginWorker (action: AnyAction) {
 
 function * loginUWorker (action: AnyAction) {
   try {
-    const userInfoUrl = `${AUTH_URL}/userinfo`
     const userinfo = yield call(request, {
-      url: userInfoUrl,
+      url: `${AUTH_URL}/userinfo`,
       method: 'GET',
       headers: {
         'Authorization': `Bearer ${action.payload.token}`,
@@ -58,15 +74,63 @@ function * loginUWorker (action: AnyAction) {
     const userId = user.id
 
     yield put(authActions.loginUserSuccess({
-      user: { fName: userName[0], lName: userName[userName.length - 1], id: userId },
+      user: {
+        fName: userName[0],
+        lName: userName[userName.length - 1],
+        id: userId,
+        role: {
+          id: user.role_id,
+          name: roleNameOptions[user.role_id - 1],
+        },
+      },
     }))
   } catch (error) {
     yield put(authActions.loginUserError(error))
   }
 }
 
+function * forgotPasswordSaga (action: AnyAction) {
+  try {
+    yield call(request, {
+      url: `${API_URL}/users/forgot`,
+      method: 'POST',
+      headers: {
+        'Content-Type': 'application/json',
+      },
+      data: JSON.stringify(action.payload),
+    })
+
+    yield put(authActions.forgotPasswordSuccess())
+  } catch (error) {
+    // TODO: change to action
+    authActions.forgotPasswordError(error)
+  }
+}
+
+function * recoverPasswordSaga (action: AnyAction) {
+  try {
+    yield call(request, {
+      url: `${API_URL}/users/recover`,
+      method: 'POST',
+      headers: {
+        'Content-Type': 'application/json',
+      },
+      data: JSON.stringify(action.payload),
+    })
+
+    yield put(authActions.recoverPasswordSuccess())
+    action.history.replace('auth/login')
+    yield put(openNotification('success', 'Password changed, you can login.', 3000))
+  } catch (error) {
+    // TODO: change to action
+    authActions.recoverPasswordError(error)
+  }
+}
+
 export function * rootSaga () {
   yield takeLatest(LOGIN, loginWorker)
   yield takeLatest([LOGIN_SUCCESS, LOGIN_USER], loginUWorker)
+  yield takeLatest(FORGOT_PASSWORD, forgotPasswordSaga)
+  yield takeLatest(RECOVER_PASSWORD, recoverPasswordSaga)
 }
 export default rootSaga
