@@ -11,46 +11,42 @@ const slideOvalUrl = requireImage('slide-oval.png')
  */
 export const slidesConfig: SlideConfig[] = [
   {
-    key: 'slide-exp',
     title: 'sandboxPage.slides.one.title',
-    p1: 'sandboxPage.slides.one.p1',
-    p2: 'sandboxPage.slides.one.p2',
-    btnStr: 'sandboxPage.slides.one.btnStr',
-    linkTo: '/auth/register',
+    subtitle: 'sandboxPage.slides.one.subtitle',
+    content: 'sandboxPage.slides.one.content',
+    button: 'sandboxPage.slides.one.button',
+    link: '/auth/register',
     btn: 1,
-    imgUrl: experienceTheSandboxUrl,
+    img: experienceTheSandboxUrl,
   },
   {
-    key: 'slide-proof',
     title: 'sandboxPage.slides.two.title',
-    p1: 'sandboxPage.slides.two.p1',
-    p2: 'sandboxPage.slides.two.p2',
-    btnStr: 'sandboxPage.slides.two.btnStr',
-    linkTo: '/',
+    subtitle: 'sandboxPage.slides.two.subtitle',
+    content: 'sandboxPage.slides.two.content',
+    button: 'sandboxPage.slides.two.button',
+    link: '/',
     disabled: true,
     btn: 2,
-    imgUrl: securePocUrl,
+    img: securePocUrl,
   },
   {
-    key: 'slide-features',
     title: 'sandboxPage.slides.three.title',
-    p1: 'sandboxPage.slides.three.p1',
-    p2: 'sandboxPage.slides.three.p2',
-    btnStr: 'sandboxPage.slides.three.btnStr',
-    linkTo: '/',
+    subtitle: 'sandboxPage.slides.three.subtitle',
+    content: 'sandboxPage.slides.three.content',
+    button: 'sandboxPage.slides.three.button',
+    link: '/',
     disabled: true,
     btn: 3,
-    imgUrl: sailTheSeasUrl,
+    img: sailTheSeasUrl,
   },
   {
-    key: 'slide-docs',
     title: 'sandboxPage.slides.four.title',
-    p1: 'sandboxPage.slides.four.p1',
-    btnStr: 'sandboxPage.slides.four.btnStr',
-    linkTo: '/',
+    subtitle: 'sandboxPage.slides.four.subtitle',
+    button: 'sandboxPage.slides.four.button',
+    link: '/',
     disabled: true,
     btn: 2,
-    imgUrl: slideOvalUrl,
+    img: slideOvalUrl,
   },
 ]
 
