@@ -6,12 +6,13 @@ import { combineReducers } from 'redux'
 import { connectRouter } from 'connected-react-router'
 import { History } from 'history'
 
-import register from 'components/RegisterPortal/ducks'
+import register from 'components/RegisterForm/ducks'
 import informDialog from 'components/InformDialog/ducks'
 import auth from 'containers/Auth/ducks'
 import applications from 'containers/Applications/ducks'
 import subscriptions from 'containers/Subscriptions/ducks'
 import notifications from 'containers/NotificationStack/ducks'
+import profile from 'containers/Profile/ducks'
 
 export default (history: History<any>) => combineReducers({
   router: connectRouter(history),
@@ -21,4 +22,5 @@ export default (history: History<any>) => combineReducers({
   subscriptions,
   informDialog,
   notifications,
+  profile,
 })
