@@ -3,10 +3,11 @@
  */
 
 import auth from 'containers/Auth/sagas'
-import register from 'components/RegisterPortal/sagas'
+import register from 'components/RegisterForm/sagas'
 import informDialog from 'components/InformDialog/sagas'
 import applications from 'containers/Applications/sagas'
 import subscriptions from 'containers/Subscriptions/sagas'
+import profile from 'containers/Profile/sagas'
 
 const sagas = [
   auth,
@@ -14,6 +15,7 @@ const sagas = [
   applications,
   subscriptions,
   informDialog,
+  profile,
 ]
 
 export default sagas
