@@ -14,6 +14,11 @@ export interface ErrorReason {
   message: string,
 }
 
+export type RequestStatus = {
+  isRequesting: boolean,
+  error: string,
+}
+
 function checkToken (response: AxiosResponse) {
   const search = response.request.responseURL.split('?')[1]
   const urlParams = new URLSearchParams(search)
@@ -26,7 +31,8 @@ function checkToken (response: AxiosResponse) {
 }
 
 async function checkStatus (response: AxiosResponse) {
-  if (response.statusText === 'OK') {
+  // TODO: we should review this reponse handling
+  if (response.statusText === 'OK' || response.statusText === 'Created') {
     // TODO add this back when API changes
     // const contentType = response.headers.get('Content-Type')
 
